1. Description：VulkanPipelineState，解决编译错误，compile ok 2. Label:[feat,fix,revert,recommit]feat 3. ImpactBetweenModules:NA/不涉及 4. ImpactOnPerformance:NA/不涉及 5. CaseId:NA/不涉及 6. Others:NA/不涉及

+++ Source/Platform/Vulkan/VulkanPipelineState.cpp
@@ -87,8 +87,8 @@ namespace MonsterRender::RHI::Vulkan {
     uint32 VulkanPipelineState::getSize() const {
         // Estimate pipeline size based on shader modules and state
         uint32 size = 0;
-        size += m_shaderModules.size() * sizeof(VkShaderModule);
-        size += m_reflectionData.size() * sizeof(ShaderReflectionData);
+        size += static_cast<uint32>(m_shaderModules.size()) * sizeof(VkShaderModule);
+        size += static_cast<uint32>(m_reflectionData.size()) * sizeof(ShaderReflectionData);
         size += sizeof(VkPipeline) + sizeof(VkPipelineLayout) + sizeof(VkRenderPass);
         return size;
     }
